requirements.txt
aiohttp==3.9.3
aioredis==1.3.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.61.2
fastapi-cache2==0.2.1
frozenlist==1.4.1
h11==0.14.0
hiredis==2.3.2
idna==3.6
iniconfig==2.0.0
multidict==6.0.5
packaging==23.2
pendulum==3.0.0
pluggy==1.4.0
pydantic==1.10.14
pydantic_core==2.16.2
pytest==8.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.13.6
time-machine==2.13.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
yarl==1.9.4
